[6803] Extract explorer selection logic into a dedicated hook
Bug: https://github.com/eclipse-sirius/sirius-web/issues/6803
