+++ packages/sirius-web/frontend/sirius-web-application/src/views/edit-project/workbench-views/explorer/ExplorerView.tsx
@@ -12,34 +12,27 @@
  *******************************************************************************/
 import {
   RepresentationLoadingIndicator,
-  Selection,
-  SelectionEntry,
-  useSelection,
   WorkbenchViewComponentProps,
   WorkbenchViewHandle,
 } from '@eclipse-sirius/sirius-components-core';
 import {
   FilterBar,
-  GQLGetTreePathVariables,
-  GQLTree,
-  GQLTreeItem,
   TreeToolBar,
   TreeToolBarContext,
   TreeToolBarContextValue,
   TreeView,
-  useTreePath,
-  useTreeSelection,
 } from '@eclipse-sirius/sirius-components-trees';
 import AccountTreeIcon from '@mui/icons-material/AccountTree';
 import Box from '@mui/material/Box';
 import { Theme } from '@mui/material/styles';
 import Typography from '@mui/material/Typography';
-import { ForwardedRef, forwardRef, useCallback, useContext, useEffect, useRef, useState } from 'react';
+import { ForwardedRef, forwardRef, useContext, useEffect, useRef, useState } from 'react';
 import { makeStyles } from 'tss-react/mui';
 import { DuplicateObjectKeyboardShortcut } from './context-menu-contributions/duplicate-object/DuplicateObjectKeyboardShortcut';
 import { ExplorerViewConfiguration, ExplorerViewState } from './ExplorerView.types';
 import { TreeDescriptionsMenu } from './TreeDescriptionsMenu';
 import { useExplorerDescriptions } from './useExplorerDescriptions';
+import { useExplorerSelection } from './useExplorerSelection';
 import { useExplorerSubscription } from './useExplorerSubscription';
 import { GQLTreeEventPayload, GQLTreeRefreshedEventPayload } from './useExplorerSubscription.types';
 import { useExplorerViewHandle } from './useExplorerViewHandle';
@@ -75,37 +68,15 @@ export const ExplorerView = forwardRef<WorkbenchViewHandle, WorkbenchViewCompone
     const initialExplorerViewConfiguration: ExplorerViewConfiguration =
       initialConfiguration as unknown as ExplorerViewConfiguration;
 
+    const configuredActiveTreeDescriptionId = initialExplorerViewConfiguration?.activeTreeDescriptionId ?? null;
+
     const [state, setState] = useState<ExplorerViewState>({
       filterBar: false,
       filterBarText: '',
       filterBarTreeFiltering: false,
       tree: null,
-      selectedTreeItemIds: [],
-      singleTreeItemSelected: null,
-      selectionPivotTreeItemId: null,
     });
-
-    // If we are requested to reveal the global selection, we need to compute the tree path to expand
-    const { getTreePath, data: treePathData } = useTreePath();
-
-    const applySelection = (selection: Selection) => {
-      const newSelectedTreeItemIds = selection.entries.map((entry) => entry.id);
-      setState((prevState) => ({
-        ...prevState,
-        selectedTreeItemIds: newSelectedTreeItemIds,
-      }));
-
-      if (state.tree && newSelectedTreeItemIds.length > 0) {
-        const variables: GQLGetTreePathVariables = {
-          editingContextId,
-          treeId: state.tree.id,
-          selectionEntryIds: newSelectedTreeItemIds,
-        };
-        getTreePath({ variables });
-      }
-    };
-
-    const configuredActiveTreeDescriptionId = initialExplorerViewConfiguration?.activeTreeDescriptionId ?? null;
+    const treeId: string | null = state.tree?.id || null;
 
     const { explorerDescriptions } = useExplorerDescriptions(editingContextId);
     const { activeTreeDescriptionId, expanded, maxDepth, onExpandedElementChange, setActiveDescriptionId } =
@@ -121,7 +92,16 @@ export const ExplorerView = forwardRef<WorkbenchViewHandle, WorkbenchViewCompone
       initialExplorerViewConfiguration?.activeTreeFilters ?? []
     );
 
-    useExplorerViewHandle(id, state.tree?.id, treeFilters, activeTreeDescriptionId, applySelection, ref);
+    const {
+      selectedTreeItemIds,
+      singleTreeItemSelected,
+      onRevealSelection,
+      onTreeItemClick,
+      applySelection,
+      setSelectedTreeItemIds,
+    } = useExplorerSelection(editingContextId, treeId, expanded, onExpandedElementChange);
+
+    useExplorerViewHandle(id, treeId, treeFilters, activeTreeDescriptionId, applySelection, ref);
 
     const treeToolBarContributionComponents = useContext<TreeToolBarContextValue>(TreeToolBarContext).map(
       (contribution) => contribution.props.component
@@ -168,41 +148,6 @@ export const ExplorerView = forwardRef<WorkbenchViewHandle, WorkbenchViewCompone
       return null;
     }, [treeElement]);
 
-    const { selection, setSelection } = useSelection();
-    const { treeItemClick } = useTreeSelection();
-
-    const selectionKey: string = selection?.entries
-      .map((entry) => entry.id)
-      .sort()
-      .join(':');
-
-    const revealSelection = useCallback(() => {
-      if (state.tree && selection.entries.length > 0) {
-        const variables: GQLGetTreePathVariables = {
-          editingContextId,
-          treeId: state.tree.id,
-          selectionEntryIds: selection.entries.map((entry) => entry.id),
-        };
-        getTreePath({ variables });
-      }
-    }, [editingContextId, selectionKey, state.tree, getTreePath]);
-
-    useEffect(() => {
-      if (treePathData && treePathData.viewer?.editingContext?.treePath) {
-        const expandedIds = treePathData.viewer.editingContext.treePath.treeItemIdsToExpand.filter(
-          (id) => !expanded.includes(id)
-        );
-        const newExpandedIds = [...expanded, ...expandedIds];
-        onExpandedElementChange(newExpandedIds, treePathData.viewer.editingContext.treePath.maxDepth);
-        setState((prevState) => {
-          return {
-            ...prevState,
-            selectedTreeItemIds: selection.entries.map((entry) => entry.id),
-          };
-        });
-      }
-    }, [treePathData]);
-
     let filterBar: JSX.Element = <div />;
     if (state.filterBar) {
       filterBar = (
@@ -231,23 +176,6 @@ export const ExplorerView = forwardRef<WorkbenchViewHandle, WorkbenchViewCompone
         </div>
       );
     }
-
-    const onTreeItemClick = (event: React.MouseEvent<HTMLDivElement, MouseEvent>, tree: GQLTree, item: GQLTreeItem) => {
-      var localSelection = treeItemClick(event, tree, item, state.selectedTreeItemIds, true);
-      setState((prevState) => ({
-        ...prevState,
-        selectedTreeItemIds: localSelection.selectedTreeItemIds,
-        singleTreeItemSelected: localSelection.singleTreeItemSelected,
-      }));
-      var globalSelection = treeItemClick(
-        event,
-        state.tree,
-        item,
-        selection.entries.map((entry) => entry.id),
-        true
-      );
-      setSelection({ entries: globalSelection.selectedTreeItemIds.map<SelectionEntry>((id) => ({ id })) });
-    };
 
     const treeDescriptionSelector: JSX.Element = explorerDescriptions.length > 1 && (
       <TreeDescriptionsMenu
@@ -292,7 +220,7 @@ export const ExplorerView = forwardRef<WorkbenchViewHandle, WorkbenchViewCompone
                 editingContextId={editingContextId}
                 readOnly={readOnly}
                 treeFilters={treeFilters}
-                onRevealSelection={revealSelection}
+                onRevealSelection={onRevealSelection}
                 onTreeFilterMenuItemClick={setTreeFilters}
                 onFilter={() => {
                   setState((prevState) => {
@@ -308,12 +236,8 @@ export const ExplorerView = forwardRef<WorkbenchViewHandle, WorkbenchViewCompone
                 target={treeElement?.current}
                 editingContextId={editingContextId}
                 readOnly={readOnly}
-                selectedTreeItem={state.singleTreeItemSelected}
-                selectTreeItems={(selectedTreeItemIds: string[]) =>
-                  setState((prevState) => {
-                    return { ...prevState, selectedTreeItemIds };
-                  })
-                }>
+                selectedTreeItem={singleTreeItemSelected}
+                selectTreeItems={setSelectedTreeItemIds}>
                 {filterBar}
                 <div className={styles.treeContent}>
                   <TreeView
@@ -326,12 +250,8 @@ export const ExplorerView = forwardRef<WorkbenchViewHandle, WorkbenchViewCompone
                     expanded={expanded}
                     maxDepth={maxDepth}
                     onTreeItemClick={onTreeItemClick}
-                    selectTreeItems={(selectedTreeItemIds: string[]) =>
-                      setState((prevState) => {
-                        return { ...prevState, selectedTreeItemIds };
-                      })
-                    }
-                    selectedTreeItemIds={state.selectedTreeItemIds}
+                    selectTreeItems={setSelectedTreeItemIds}
+                    selectedTreeItemIds={selectedTreeItemIds}
                     data-testid="explorer://"
                     useTreePalette={state.tree.capabilities.useTreePalette}
                   />
